requirements.txt
# Langchain Core Components (Minimum versions - Attempt 4)
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
langchain-google-genai>=2.0.0
langchain-mcp-adapters>=0.0.7 # Keep this specific for now

# --- Other packages ---
google-api-python-client>=2.88.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0